acbgen using raw module
